try to use int64 to handle larger values

+++ math.go
@@ -1,7 +1,7 @@
 package main
 
-func factor(num int, result *[]int) {
-	for a := 2; a*a <= num; a++ {
+func factor(num int64, result *[]int64) {
+	for a := int64(2); a*a <= num; a++ {
 		for num%a == 0 {
 			*result = append(*result, a)
 			num /= a
@@ -12,8 +12,8 @@ func factor(num int, result *[]int) {
 	}
 }
 
-func prime(num int) bool {
-	for a := 2; a*a <= num; a++ {
+func prime(num int64) bool {
+	for a := int64(2); a*a <= num; a++ {
 		if num%a == 0 {
 			return false
 		}
